@@ -12,46 +12,47 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^13.3.11",
-    "@angular/cdk": "^13.3.9",
-    "@angular/common": "^13.3.11",
-    "@angular/compiler": "^13.3.11",
-    "@angular/core": "^13.3.11",
+    "@angular/animations": "^ 14.1.3",
+    "@angular/cdk": "^ 14.1.3",
+    "@angular/common": "^ 14.1.3",
+    "@angular/compiler": "^ 14.1.3",
+    "@angular/core": "^ 14.1.3",
     "@angular/flex-layout": "^12.0.0-beta.35",
-    "@angular/forms": "^13.3.11",
-    "@angular/material": "^13.3.9",
-    "@angular/platform-browser": "^13.3.11",
-    "@angular/platform-browser-dynamic": "^13.3.11",
-    "@angular/router": "^13.3.11",
-    "@ngrx/component-store": "^13.2.0",
-    "@ngrx/effects": "^13.2.0",
-    "@ngrx/entity": "^13.2.0",
-    "@ngrx/router-store": "^13.2.0",
-    "@ngrx/store": "^13.2.0",
-    "@ngrx/store-devtools": "^13.2.0",
+    "@angular/forms": "^ 14.1.3",
+    "@angular/material": "^ 14.1.3",
+    "@angular/platform-browser": "^ 14.1.3",
+    "@angular/platform-browser-dynamic": "^ 14.1.3",
+    "@angular/router": "^ 14.1.3",
+    "@ngrx/component-store": "^14.2.0",
+    "@ngrx/effects": "^14.2.0",
+    "@ngrx/entity": "^14.2.0",
+    "@ngrx/router-store": "^14.2.0",
+    "@ngrx/store": "^14.2.0",
+    "@ngrx/store-devtools": "^14.2.0",
+    "@vime/angular": "^5.3.1",
+    "@vime/core": "^5.3.1",
+    "cookie-parser": "^1.4.6",
     "cors": "^2.8.5",
     "express": "^4.18.1",
     "helmet": "^6.0.0",
     "http-proxy-middleware": "^2.0.6",
+    "morgan": "^1.10.0",
     "ngx-infinite-scroll": "^14.0.0",
     "ngx-ui-loader": "^13.0.0",
     "nodemon": "^2.0.19",
     "normalize.css": "^8.0.1",
-    "highlight.js": "^11.6.0",
-    "simplebar": "~5.3.8",
     "rxjs": "~7.5.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.11.4"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^13.3.9",
-    "@angular/cli": "~13.3.9",
-    "@angular/compiler-cli": "^13.3.11",
-    "@ngrx/schematics": "^13.2.0",
+    "@angular-devkit/build-angular": "^14.1.3",
+    "@angular/cli": "~14.1.3",
+    "@angular/compiler-cli": "^14.1.3",
+    "@ngrx/schematics": "^14.2.0",
     "@types/express": "^4.17.13",
     "@types/jasmine": "~4.0.0",
     "@types/node": "^18.7.13",
-    "axios": "^0.27.2",
     "https-proxy-agent": "^5.0.1",
     "jasmine-core": "~4.2.0",
     "karma": "~6.4.0",
