@@ -1,10 +1,10 @@
-import { Injectable } from '@angular/core';
-import { Actions, createEffect, ofType } from '@ngrx/effects';
-import {catchError, map, concatMap, debounceTime, switchMap, skip, takeUntil} from 'rxjs/operators';
-import { EMPTY, of, asyncScheduler} from 'rxjs';
-import {ArtistsApiActions, SearchArtistPageActions} from "../actions";
+import {Injectable} from '@angular/core';
+import {Actions, createEffect, ofType} from '@ngrx/effects';
+import {catchError, map, debounceTime, switchMap, skip, takeUntil, tap} from 'rxjs/operators';
+import {EMPTY, of, asyncScheduler} from 'rxjs';
+import { ArtistsApiActions, SearchArtistPageActions} from "../actions";
 import {ApiDataService} from "../../core/services";
-import {ArtistSearchResult} from "../models";
+import {ActivatedRoute, Router} from "@angular/router";
 
 
 @Injectable()
@@ -12,11 +12,11 @@ export class ArtistsEffects {
 
   search$ = createEffect(
     () =>
-      ({ debounce = 800, scheduler = asyncScheduler } = {}) =>
+      ({debounce = 1000, scheduler = asyncScheduler} = {}) =>
         this.actions$.pipe(
           ofType(SearchArtistPageActions.searchArtists),
           debounceTime(debounce, scheduler),
-          switchMap(({ query }) => {
+          switchMap(({query, resourceName}) => {
             if (query === '') {
               return EMPTY;
             }
@@ -28,13 +28,28 @@ export class ArtistsEffects {
 
             return this.apiDataService.searchArtists(query).pipe(
               takeUntil(nextSearch$),
-              map((resp: any) =>{
+              map((resp: any) => {
                 console.log(resp.data)
                 const data = resp.data
-                return   ArtistsApiActions.searchSuccess({ data })}),
+                const url = resp.next
+                return ArtistsApiActions.searchSuccess({data})
+              }),
               catchError((err) =>
-                of(ArtistsApiActions.searchFailure({ errorMsg: err.message }))
-              )
+                of(ArtistsApiActions.searchFailure({errorMsg: err.message}))
+              ),
+              tap(() => {
+                return this.router
+                  .navigate(['/artists'], {
+                    relativeTo: this.route,
+                    queryParams: {
+                      q: query,
+
+                    },
+                    queryParamsHandling: 'merge',
+                  })
+                  .then();
+
+              })
             );
           })
         )
@@ -43,6 +58,9 @@ export class ArtistsEffects {
 
   constructor(
     private actions$: Actions,
-    private apiDataService: ApiDataService
-  ) {}
+    private apiDataService: ApiDataService,
+    private router: Router,
+    private route: ActivatedRoute
+  ) {
+  }
 }
