@@ -1,14 +1,12 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
-import {MyArtistsCollectionPageComponent} from "./containers/my-artists-collection-page.component";
 import {ViewArtistPageComponent} from "./containers/view-artist-page.component";
 import {SearchArtistsPageComponent} from "./containers/search-artists-page.component";
 import {ArtistExistsGuard} from "./guards";
 
 const routes: Routes = [
-
   {
-    path: 'search',
+    path: '',
     component: SearchArtistsPageComponent,
     data: {title: 'Deezer Artists'},
   },
@@ -17,11 +15,8 @@ const routes: Routes = [
     component: ViewArtistPageComponent,
     canActivate: [ArtistExistsGuard],
     data: {title: 'Artist Details'}
-  }, {
-    path: '',
-    component: SearchArtistsPageComponent,
-    data: {title: 'My Artists Collection'}
   },
+
   ]
 
 @NgModule({
