@@ -3,39 +3,70 @@ import {Album, Artist, ArtistSearchResult, ArtistTopTrack} from "../models";
 import {ArtistAlbumsStore} from "../component-store/artist-albums.store";
 import {Observable} from "rxjs";
 import {ArtistTopTracksStore} from "../component-store/artist-top-tracks.store";
+import {MediaObserver} from "@angular/flex-layout";
 
 @Component({
   selector: 'app-artist-detail',
   template: `
     <mat-card>
-      <div>
-        <div class="artist-details">
-          <mat-card class="example-card">
-            <mat-card-header>
-              <mat-card-title>{{artist.name}}</mat-card-title>
-            </mat-card-header>
-            <img mat-card-image [src]="artist.picture_medium" alt="Artist_Image">
+      <div class="toptracks-container">
+        <mat-card class="artist-details" fxLayout="row" fxLayoutAlign="space-between stretch">
+          <mat-card fxFlex.xs>
+          <img mat-card-image [src]="artist.picture_medium" alt="Artist_Image">
             <mat-card-content>
-              <h3>
+              <div fxLayout="column" fxLayoutAlign="space-between start" >
+              <span>{{artist.name}}</span>
+              <span>
                 <strong>
-                {{artist.nb_fan | convertFans}}
-                </strong>fans</h3>
+                  {{artist.nb_fan | convertFans}}
+                </strong>fans</span>
+              </div>
             </mat-card-content>
           </mat-card>
 
-        </div>
-        <div class="artist-top-tracks">
-          <app-artist-top-tracks [topTracks]="topTracks$"></app-artist-top-tracks>
-        </div>
+          <div class="artist-top-tracks" [hidden]="media.isActive('xs')" fxFlex>
+            <app-artist-top-tracks [topTracks]="topTracks$"></app-artist-top-tracks>
+          </div>
 
-        <div class="artist-albums">
+        </mat-card>
+
+      </div>
+
+      <div class="artist-album">
           <app-artist-albums [albums]="albums$"></app-artist-albums>
-        </div>
-
       </div>
     </mat-card>
   `,
-  styles: [],
+  styles: [`
+    :host {
+      //display: flex;
+      //justify-content: center;
+    }
+    .artist-top-tracks{
+      margin: 8px;
+      span {
+        //overflow: hidden;
+        width: 100%;
+        text-overflow: ellipsis;
+        white-space: nowrap;
+      }
+
+
+    }
+    .artist-album{
+      margin-top: 21px;
+    }
+
+
+      mat-card img {
+        max-width: 400px;
+        max-height: 400px;
+      }
+
+    .artist-details {
+      margin-top: 60px;
+    }
+  `],
   changeDetection: ChangeDetectionStrategy.OnPush,
   providers: [ArtistAlbumsStore, ArtistTopTracksStore]
 })
@@ -52,7 +83,9 @@ export class ArtistDetailComponent implements OnInit {
 
   constructor(
     private readonly artistAlbumsStore: ArtistAlbumsStore,
-    private readonly artistTopTracksStore: ArtistTopTracksStore
+    private readonly artistTopTracksStore: ArtistTopTracksStore,
+    public media: MediaObserver,
+
   ) {
   }
 
