@@ -1,4 +1,4 @@
-import { ChangeDetectionStrategy, Component } from '@angular/core';
+import {ChangeDetectionStrategy, Component} from '@angular/core';
 import {Observable} from "rxjs";
 import {Store} from "@ngrx/store";
 import * as fromArtists from '../reducers';
@@ -7,9 +7,17 @@ import {ArtistSearchResult} from "../models";
 @Component({
   selector: 'app-search-artists',
   template: `
-    <app-artist-preview-list [artists]="(searchResults$ | async)!"> </app-artist-preview-list>
+    <div class="multiview-container">
+      <app-artist-preview-list [artists]="(searchResults$ | async)!"> </app-artist-preview-list>
+    </div>
   `,
   styles: [
+    `.multiview-container {
+      width: 100%;
+      height: 100vh;
+      overflow: scroll;
+    }
+    `
   ],
   changeDetection: ChangeDetectionStrategy.OnPush
 })
@@ -17,10 +25,10 @@ export class SearchArtistsPageComponent {
   searchResults$: Observable<ArtistSearchResult[]>;
 
   constructor(
-    private store: Store
-
+    private store: Store,
   ) {
     this.searchResults$ = store.select(fromArtists.selectSearchResults)
  }
+
 
 }
