@@ -7,11 +7,7 @@ import {ApiDataService} from "../../core/services";
 import * as fromArtists from '../../artists/reducers'
 import {ArtistActions} from "../actions";
 
-/**
- * Guards are hooks into the route resolution process, providing an opportunity
- * to inform the router's navigation process whether the route should continue
- * to activate this route. Guards must return an observable of true or false.
- */
+
 @Injectable({
   providedIn: 'root',
 })
@@ -22,11 +18,7 @@ export class ArtistExistsGuard implements CanActivate {
     private router: Router
   ) {}
 
-  /**
-   * This method creates an observable that waits for the `loaded` property
-   * of the collection state to turn `true`, emitting one time once loading
-   * has finished.
-   */
+
   waitForCollectionToLoad(): Observable<boolean> {
     return this.store.select(fromArtists.selectCollectionLoaded).pipe(
       filter((loaded) => loaded),
@@ -34,10 +26,7 @@ export class ArtistExistsGuard implements CanActivate {
     );
   }
 
-  /**
-   * This method checks if a book with the given ID is already registered
-   * in the Store
-   */
+
   hasArtistInStore(id: string): Observable<boolean> {
     return this.store.select(fromArtists.selectArtistEntities).pipe(
       map((entities) => !!entities[id]),
@@ -45,10 +34,6 @@ export class ArtistExistsGuard implements CanActivate {
     );
   }
 
-  /**
-   * This method loads a book with the given ID from the API and caches
-   * it in the store, returning `true` or `false` if it was found.
-   */
   hasArtistInApi(id: string): Observable<boolean> {
     return this.apiDataService.getArtist(id).pipe(
       map((ArtistEntity) => ArtistActions.loadArtist({ artist: ArtistEntity })),
@@ -61,11 +46,7 @@ export class ArtistExistsGuard implements CanActivate {
     );
   }
 
-  /**
-   * `hasBook` composes `hasBookInStore` and `hasBookInApi`. It first checks
-   * if the book is in store, and if not it then checks if it is in the
-   * API.
-   */
+
   hasArtist(id: string): Observable<boolean> {
     return this.hasArtistInStore(id).pipe(
       switchMap((inStore) => {
@@ -77,23 +58,8 @@ export class ArtistExistsGuard implements CanActivate {
     );
   }
 
-  /**
-   * This is the actual method the router will call when our guard is run.
-   *
-   * Our guard waits for the collection to load, then it checks if we need
-   * to request a book from the API or if we already have it in our cache.
-   * If it finds it in the cache or in the API, it returns an Observable
-   * of `true` and the route is rendered successfully.
-   *
-   * If it was unable to find it in our cache or in the API, this guard
-   * will return an Observable of `false`, causing the router to move
-   * on to the next candidate route. In this case, it will move on
-   * to the 404 page.
-   */
+
   canActivate(route: ActivatedRouteSnapshot): Observable<boolean> {
-    // return this.waitForCollectionToLoad()!!.pipe(
-    //   switchMap(() => this.hasArtist(route.params['id']))
-    // );
     return this.hasArtist(route.params['id'])
   }
 }
