@@ -1,25 +1,32 @@
-import { Component } from '@angular/core';
+import {Component} from '@angular/core';
+import {routeAnimations} from "../../../shared/animations/route.animations";
 
 @Component({
   selector: 'app-layout',
   template: `
-    <mat-sidenav-container fullscreen>
-      <ng-content></ng-content>
-    </mat-sidenav-container>
+    <mat-drawer-container  [@routeAnimations]="
+        outlet && outlet.isActivated && outlet.activatedRouteData" fullscreen>
+      <mat-drawer-content>
+        <ng-content></ng-content>
+      <router-outlet #outlet="outlet"></router-outlet>
+      </mat-drawer-content>
+    </mat-drawer-container>
   `,
   styles: [
     `
-      mat-sidenav-container {
-        margin-top: 65px;
-
-        background: rgba(0, 0, 0, 0.03);
-      }
-
-      :host ::ng-deep  {
+      :host ::ng-deep {
         -webkit-font-smoothing: antialiased;
         -moz-osx-font-smoothing: grayscale;
       }
+      .mat-drawer-container{
+        position: unset;
+
+
+      }
     `,
   ],
+  animations: [routeAnimations],
+
+
 })
 export class LayoutComponent {}
