@@ -1,7 +1,7 @@
 import { Component, ChangeDetectionStrategy } from '@angular/core';
 
 @Component({
-  selector: 'bc-not-found-page',
+  selector: 'app-not-found-page',
   changeDetection: ChangeDetectionStrategy.OnPush,
   template: `
     <mat-card>
@@ -10,9 +10,8 @@ import { Component, ChangeDetectionStrategy } from '@angular/core';
         <p>Hey! It looks like this page doesn't exist yet.</p>
       </mat-card-content>
       <mat-card-actions>
-        <button mat-raised-button color="primary" routerLink="/">
-          Take Me Home
-        </button>
+        <button mat-raised-button color="primary" routerLink="/artists">
+          Take Me Home</button>
       </mat-card-actions>
     </mat-card>
   `,
