@@ -1,5 +1,7 @@
 export const environment = {
   production: true,
-  Deezer_API:'https://api.deezer.com'
+  Deezer_API:'https://api.deezer.com',
+  Deezer_API_LOCAL:'http://localhost:3000/api',
+  ngrx_logs: false,
 
 };
